vimrc: disable rope completion on dot
This hasn't worked out so well, I usually seem to know what I want to
type and the completion often hangs for too long.

Also it has the unwelcome side-effect of automatically opening a
documentation window at the top of the screen. It's possible to call
this up with <C-c>d if desired.